requirements.txt
pytest-playwright
openai
